Clean up api aread with actual used data
