+++ pages/api/users/[id]/metrics/[day].js
@@ -2,17 +2,31 @@
 import prisma from '../../../../../lib/prisma';
 
 export default async function handler(req, res) {
-  if (req.method === 'PUT' || req.method === 'POST') {
+  // (Format is MM-DD-YYYY)
+  const date_id = req.query.day;
+  const user_id = Number(req.query.id);
+  let theDate = new Date(date_id);
+  let theDateStr = theDate.toISOString();
+
+  if (req.method === 'GET') {
     const id = Number(req.query.id);
     const eid = Number(req.query.eid);
     const entries = await prisma.User_metric_data.findMany({
       where: {
-        id: eid,
         user_id: id,
+        date: {
+          lte: theDateStr,
+          gte: new Date(new Date().setDate(theDate.getDate() - 30)),
+        }
       },
       include: {
         metrics: true,
-      }
+      },
+      orderBy: [
+        {
+          date: 'desc'
+        }
+      ],
     });
     return await res.status(200).json(entries);
   } else {
